fix issue #2
The heuristics for initial parameter estimates used to rely explicitly
on the first (few) empirical MSD data point(s). This fails if a
trajectory does not have any directly adjacent frames, such as the
example in issue #2.

Fixed by using the first *valid* empirical MSD point in place of the
"real" first one.