check parameter feasibility (bounds) in Fit._penalty
Doing this check manually means we can enforce bounds on fixed
parameters.

We only check feasibility; if a soft approach to the bound is needed,
implement a constraint. Alternatively, one might be tempted to implement
a generic soft approach to the bound. This does not make sense:
 + we would have to choose a generally valid notion of "closeness"; that
might still be fine (e.g. use linearization distance of 1e-7).
 + some bounds (like 0 for anomalous scaling) are very much attainable
and are attained during fitting. We do not want to generically mess with
the likelihood close to these bounds.