Fix issue #3
This issue was two-fold: exceptions from cython code would not get
raised but skipped; and NPXFit would raise exceptions during Cholesky
decomposition, even though the covariance should always be positive
definite in this case.

For the MWE given in the issue, this problem can actually be solved by
just addressing the first issue: telling the C function _core_logL to
forward exceptions through an error return value lets the rest of the
code take care of these scenarios.

To be on the safe side, we also decrease the upper bound for the
exponent (this is what caused the numerical issues with non
positive-definite covariance matrix) from 2-1e10 to 1.99. For any
practical purposes this should still be fine, and numerically hopefully
more stable.