(slightly) update calculation of C
Line profiling indicated that assembling the covariance matrix was the
biggest performance bottleneck, so I tried to optimize. It never did
much. Turns out, the overhead from the line profiler distorted runtimes
more than I had expected, so in fact the bottleneck is indeed the
likelihood calculation, which is optimized already. So: nothing to be
done about this (for now).