Scale min eigenvalue in Cpositive constraint
The Fit.constraint_Cpositive compares covariance eigenvalues to some
minimum plausible value that's derived from white noise. This eigenvalue
should of course match the "units" of the covariances in the real
system, i.e. be scaled by e.g. MSD(1). This was not the case so far and
is introduced/fixed by this commit.