add adjust_prior_for_fixed_values to run_fit()
This is important, since we introduced "real" priors (i.e. normalized to
the free parameters). Since the profiler fixes a given parameter and
then runs the fit, it is important to internally still pretend that that
parameter is free, since we want to explore the likelihood landscape
associated with the original fit setting.