parallelization update
Parallelizing single likelihood evaluations (i.e. on a single dimension
of a single trajectory) was completely useless; that call is too fast to
warrant the overhead from multiprocessing.

The new scheme gives a bit more control to the user as to where (at what
level) to parallelize: Fit and Fit.evidence now have a kward
"likelihood_chunksize" that can be used to control parallelization, if
it is globally enabled by using the "noctiluca.Parallel" context
manager. By default it is set to -1 everywhere, disabling
parallelization; to enable, set it to a positive value. chunksize = 0 is
a special case that executes the whole evaluation in series, but does
shove it into a separate process; felt like it might be useful for
something.