handle BadCovarianceError
Usually, Fit should take care that it evaluates the likelihood only for
positive definite MSDs. In some edge cases, however, it is possible that
this does not work and the likelihood function tries to evaluate on a
non-positive-definite covariance matrix; meaning it will throw a
BadCovarianceError.

We used to just raise these errors, since in principle they are an
indication that something went wrong that shouldn't go wrong. But e.g.
evidence() tends to find critical edge cases like that every now and
then, so it is annoying if the whole calculation gets cancelled because
it tried to evaluate some obscure parameter combination. So the better
solution seems to be to just return Fit.max_penalty instead of raising
the error. If this behavior leads to problems in the future, we might
have to rethink.

One issue with this solution is that it make the target function for
minimization non-smooth (non-continuous even); in the case of evidence()
this would be fine, because we're evaluating on a grid. If it happens
during actual fitting, it might be more of an issue.